ENH Major vim improvements
I added several great plugins.

fugitive should largely eliminate the need to constantly open a terminal
for git.

conque may or may not turn out to be very useful, but it's sort of an
emacs-style terminal emulator inside a vim buffer.

fuzzyfinder is huge: it may or may not turn out to be a great file
browser, but (I'm hoping) that at the least it will be a wonderful
buffer browser. I added mappings for this: C-a for buffers, C-f for
files, C-d for directories. There are a number of more obscure modes for
which I did not add mappings.

solarized is not that big a deal, but it's a nice color scheme and I'm
at least giving it a try. It is pleasing, but with font as small as mine
it may not work. At the very least, I have to boldface whenever I use
it. There's a toggle for background... do :set backgound=* with * as
light or dark.

L9 is just a programming toolbox for vimscripts. Fuzzyfinder uses it.